+++ src/engine/actions/noop.ts
@@ -1,0 +1,6 @@
+import { Action } from 'engine/types'
+
+/** Do nothing this turn */
+export const NoopAction: Action = () => {
+  // noop
+}
